+++ tests/Crypto/SHA512WhirlpoolTest.php
@@ -19,45 +19,38 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-namespace OCA\UserSQL\Model;
+namespace Tests\UserSQL\Crypto;
+
+use OCA\UserSQL\Crypto\SHA512Whirlpool;
+use OCA\UserSQL\Crypto\IPasswordAlgorithm;
+use OCP\IL10N;
+use Test\TestCase;
 
 /**
- * The user entity.
+ * Unit tests for class <code>SHA512Whirlpool</code>.
  *
  * @author Marcin Łojewski <dev@mlojewski.me>
  */
-class User
+class SHA512WhirlpoolTest extends TestCase
 {
     /**
-     * @var string The UID (username).
+     * @var IPasswordAlgorithm
      */
-    public $uid;
-    /**
-     * @var string The user's email address.
-     */
-    public $email;
-    /**
-     * @var string The user's display name.
-     */
-    public $name;
-    /**
-     * @var string The user's password (hash).
-     */
-    public $password;
-    /**
-     * @var string The user's home location.
-     */
-    public $home;
-    /**
-     * @var bool Is user account active.
-     */
-    public $active;
-    /**
-     * @var bool Can user change its avatar.
-     */
-    public $avatar;
-    /**
-     * @var string The password's salt.
-     */
-    public $salt;
+    private $crypto;
+
+    public function testCheckPassword()
+    {
+        $this->assertTrue(
+            $this->crypto->checkPassword(
+                "password",
+                "a96b16ebb691dbe968b0d66d0d924cff5cf5de5e0885181d00761d87f295b2bf3d3c66187c050fc01c196ff3acaa48d3561ffd170413346e934a32280d632f2e"
+            )
+        );
+    }
+
+    protected function setUp()
+    {
+        parent::setUp();
+        $this->crypto = new SHA512Whirlpool($this->createMock(IL10N::class));
+    }
 }